+++ docs/gulpfile.js
@@ -17,13 +17,7 @@ gbm({
     build: docs.build,
     clean: docs.cleanList,
     default: ['@clean', '@build'],
-    watch: {
-      browserSync: {
-        server: 'docs/_site',
-        port: 3000,
-        reloadDelay: 0
-      }
-    }
+    watch: { browserSync: {server: 'docs/_site', port: 3000, open:false, reloadDebounce:500}}
   },
   // moduleOptions: {del: {force:true}}  // enable files outside of this project to be deleted
 });