+++ src/builders/GCSSBuilder.ts
@@ -12,18 +12,19 @@ export class GCSSBuilder extends GBuilder {
   async build() {
     this.src().chain(new CSSPlugin());
 
-    // if outfileOnly is not set, it's default value is treated to be true
+    // if outfileOnly is not set, treat the default value to be true
     if (!this.conf.outFile || this.buildOptions.outFileOnly == false) {
       //---before concat
       if (this.buildOptions.minify || this.buildOptions.minifyOnly)
-        this.pushStream().chain(GPlugin.cssnano).dest().popStream();
+        this.pushStream().chain(GPlugin.cleancss).sourceMaps().dest().popStream();
       if (!this.buildOptions.minifyOnly) this.dest();
     }
 
     // concat and the next actions
-    if (this.conf.outFile) this.chain(GPlugin.concat);
+    if (this.conf.outFile) this.chain(GPlugin.concat).sourceMaps();
     if (!this.buildOptions.minifyOnly) this.dest();
-    if (this.buildOptions.minify || this.buildOptions.minifyOnly) this.chain(GPlugin.cssnano).dest();
+    if (this.buildOptions.minify || this.buildOptions.minifyOnly)
+      this.chain(GPlugin.cleancss).sourceMaps().dest();
     return this;
   }
 }